requirements.txt
crewai
crewai-tools
python-dotenv
pdfplumber
python-docx
langchain-openai
langchain-google-genai
pydantic
streamlit
google-generativeai
